requirements.txt
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.3.0
pandas==2.2.2
tenacity==8.3.0
cloudinary==1.39.0
rich==13.7.1
